refactor: Move wasm helper to client directory and eliminate @ts-ignore
